Rename core package to landingzone
